% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build_gallery.R
\name{build_gallery}
\alias{build_gallery}
\title{Use knitr to build all the snippets in a gallery}
\usage{
build_gallery(infolder = "_R", outfolder = "snippets",
  images_dir = "images", ...)
}
\arguments{
\item{infolder}{input directory}

\item{outfolder}{directory}

\item{images_dir}{directory to store resulting images}

\item{...}{extra arguments passed on to knitr's chunk options}
}
\description{
Snippets are stored as .Rmd files in one directory: save them into
a new directory.
}

