% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean_gallery.R
\name{clean_gallery}
\alias{clean_gallery}
\title{clean out snippet output from a gallery}
\usage{
clean_gallery(gallery, infolder = "_R", outfolder = "snippets",
  images_dir = "images", caches_dir = "_caches")
}
\arguments{
\item{gallery}{directory of the gallery to be cleaned}

\item{infolder}{input directory}

\item{outfolder}{directory}

\item{images_dir}{directory to store images}

\item{caches_dir}{directory to store caches

This will not remove a .md file, image directory, or cache directory that
does not have a corresponding .Rmd file in the input (it will show a
message).}
}
\description{
Clean out the .md files, along with image and cache directories,
from an R-gallery. This doesn't remove the files in the _R directory
itself.
}

