% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmvrnorm}
\alias{cmvrnorm}
\title{cmvrnorm}
\usage{
cmvrnorm(n, mu, Sigma)
}
\arguments{
\item{n}{the number of samples required. Default 1.}

\item{mu}{a vector giving the means of variables.}

\item{Sigma}{a positive-definite symmetric specifying the covariance matrix of variabless}
}
\value{
mat a matrix of n samples of p variables on p-mvr
}
\description{
Produces one or more samples from specified multivariate normal distribution by C-plus-plus
}
\details{
Get multiple variable normal distribution via Choleski decompsition and Rcpp
}
\examples{
sigma <- matrix(c(1, 0.9, -0.3, 0.9, 1, -0.4, -0.3, -0.4, 1), ncol = 3)
mu <- c(10, 5, -3)
var(cmvrnorm(10, mu,  sigma))
}
\author{
Eric xin Zhou \url{xxz220@miami.edu}
}

