% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_gallery.R
\name{create_gallery}
\alias{create_gallery}
\title{Create a gallery in a new directory}
\usage{
create_gallery(directory, repository = "dgrtwo/rgallery-default",
  set_wd = FALSE)
}
\arguments{
\item{directory}{directory in which to create the gallery}

\item{repository}{repository to clone the initial gallery from. Either
a username/repo from github, or a full URL starting with \code{http://}}

\item{set_wd}{whether to set the working directory to the gallery afterwards}
}
\description{
This clones and sets up a new gallery directory. By default it installs
the one found at http://github.com/dgrtwo/rgallery-default.
}

