% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Timer.R
\name{Gbenchmark}
\alias{Gbenchmark}
\title{Gbenchmark}
\usage{
Gbenchmark(FUN, ...)
}
\arguments{
\item{FUN}{generation function of certain to be benchmarked}

\item{...}{other arguments passed to FUN. This is determined by FUN}
}
\value{
time A data.frame records each options time consuming
}
\description{
Gbenchmark is a function whose arguments is a simulation data generation function and its argumemts.
             Via this benchmark function, we can calculate different function's growth order with function's
             argument size changed.
}
\details{
L2 glmnet function's Gbenchmark on different dataset
}
\author{
Eric xin Zhou \url{xxz220@miami.edu}
}

