% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{modify_yaml_front_matter}
\alias{modify_yaml_front_matter}
\title{Modify yaml front matter of a file}
\usage{
modify_yaml_front_matter(infile, outfile = infile, ...)
}
\arguments{
\item{infile}{file to be modified}

\item{outfile}{file to write}

\item{...}{fields to change in front matter}
}
\description{
Modify the yaml front matter of a file. Can be used to change a file
in place. If no front matter yet exists, this creates one.
}

