% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{partition_yaml_front_matter}
\alias{partition_yaml_front_matter}
\title{Partition yaml front matter}
\usage{
partition_yaml_front_matter(input_lines)
}
\arguments{
\item{input_lines}{input lines}
}
\description{
This function comes directly from the rmarkdown package
}

