% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read.data.frame}
\alias{read.data.frame}
\title{Read an ntuple from an HDF5 ntuple file.}
\usage{
read.data.frame(hfile, groupname)
}
\arguments{
\item{hfile}{The H5File object, as created by h5file.}

\item{groupname}{character vector, the name of the Group to be read. Must be length 1.}
}
\value{
a data.frame, containing as columns the Datasets in the Group
}
\description{
Read a Group in an HDF5 ntuple file, returning a data.frame. The Datasets in the
Group will form columns in the data.frame. Datasets that are 2-dimenionsal
arrays will be broken into one column for each column in the array.
}
