% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RGallery-package.R
\docType{package}
\name{RGallery}
\alias{RGallery}
\alias{RGallery-package}
\title{RGallery: A Package for 3 questions in Stochastic Average Gradient Project}
\description{
\code{RGallery} provides a test Package for 3 questions in Stochastic Average Gradient Project
}

