% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimLR.R
\name{sim_dat_LR}
\alias{LR_simulate}
\alias{sim_dat_LR}
\title{sim_dat_LR}
\usage{
sim_dat_LR(N, p, rho = 0.95, snr = 3, nmu = 0, nsd = 1,
  control = TRUE)
}
\arguments{
\item{N}{number of observations}

\item{p}{number of predictors}

\item{rho}{correlation of X_j and X_{j'}. Default is 0.95}

\item{snr}{signal-noise ratio threahold. Default is 3.0}

\item{nmu}{noise distribution's mean. Default is 0.0}

\item{nsd}{noise distribution's sd. Default is 1.0}

\item{control}{Optional(TRUE/FALSE), whether control whiten noise and signal's SNR}
}
\value{
x p predictors' matrix for N observation.

y N observation (1 or 0) which is based on real value of y0 and noise.

beta p real beta value
}
\description{
A data simulation function which is used for glmnet benchmark. The data set size for L2-regularized
}
\details{
Logistical regression data simulation
}
\author{
Eric xin Zhou \url{xxz220@miami.edu}
}

