% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{with_wd}
\alias{with_wd}
\title{temporarily evaluate an expression in a directory}
\usage{
with_wd(dir, expr)
}
\arguments{
\item{dir}{a directory to perform an expression within}

\item{expr}{expression to evaluate}
}
\description{
Temporarily evaluate an expression in a directory, then set the directory
back to the original.
}
\details{
See here: http://plantarum.ca/code/setwd-part2/
}

