% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build_snippet.R
\name{build_snippet}
\alias{build_snippet}
\title{Knit a single .Rmd snippet}
\usage{
build_snippet(input, output, images_dir = "images", base_url = "/", ...)
}
\arguments{
\item{input}{a .Rmd file to compile}

\item{output}{the .md file to write}

\item{images_dir}{directory to save images to}

\item{base_url}{base_url of site}

\item{...}{extra arguments passed on to knitr's chunk options}
}
\description{
Knit a .Rmd snippet into a .md file. A snippet is like a "mini-vignette"
}
\details{
This makes sure in the process
that the images go into an accessible directory, and add an "image:" tag to
the front-matter displaying one of the snippet's images (to use as a
teaser on the main gallery page)

This function was originally inspired by:
http://jfisher-usgs.github.com/r/2012/07/03/knitr-jekyll/
}

